requirements.txt
fastapi
uvicorn[standard]
requests
wikipedia-api
python-dotenv
jinja2
pydantic